+++ src/main/java/com/github/yangricardo/api_spring_boot/modules/users/model/Role.java
@@ -1,7 +1,11 @@
 package com.github.yangricardo.api_spring_boot.modules.users.model;
 
 import com.github.yangricardo.api_spring_boot.shared.modules.models.BaseModel;
+
+import java.util.Collection;
+
 import javax.persistence.Entity;
+import javax.persistence.ManyToMany;
 import javax.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,18 +15,21 @@ import lombok.NoArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 
 @Entity
-@Table(name = "authorization_profiles")
+@Table(name = "roles")
 @Builder
 @Data
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
-public class AuthorizationProfile extends BaseModel implements GrantedAuthority {
+public class Role extends BaseModel implements GrantedAuthority {
 
-  private String authority;
+  private String role;
+
+  @ManyToMany(mappedBy = "roles")
+  private Collection<User> users;
 
   @Override
   public String getAuthority() {
-    return this.authority;
+    return this.role;
   }
 }